Remove :safe opt from binary_to_term call
Not 100% clear on why, since the data being stored is "inert", but `binary_to_term` with `:safe` opt fails when trying to read data after stopping and reconnecting a notebook. (Maybe DateTimes use refs or something in their internal representation?)

Regardless, it's not necessary to use `:safe` here because the data is not coming from an untrusted source--it's being read by the same program that saved it.